+++ Back/symfoAPE/src/Controller/NewsController.php
@@ -3,7 +3,10 @@
 namespace App\Controller;
 
 use App\Entity\News;
+use App\Entity\Medium;
 use App\Repository\NewsRepository;
+use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -33,9 +36,38 @@ class NewsController extends AbstractController
 
     /**
      * @Route("/api/news/create", name="news_create")
+     */
+    public function create(Request $request, EntityManagerInterface $entityManager) {
+        $newNews = json_decode($request->getContent(), true);  
+ 
+        $news = new News();
+        $news->setTitle($newNews->title);
+        $news->setContent($newNews->content);
+ 
+        $entityManager->persist($news);
+        $entityManager->flush();
+        
+        $data = [
+            'status' => 201,
+            'message' => 'Le news à été créé'
+        ];
+        return new JsonResponse($data, 201);  
+    }
+
+    /**
      * @Route("/api/news/{id}/edit", name="news_edit")
      */
-    //TODO
+    public function edit(Medium $media, Request $request) 
+    {
+        // $this->denyAccessUnlessGranted
+        $media = json_decode($request->getContent());
+        if ($media === null) {
+            return new JsonResponse(['message' =>'Ce media n\'existe pas.'], Response::HTTP_NOT_FOUND);
+        }
+
+        $response = new JsonResponse($media, 200);
+        return $response;
+    }
 
     /**
      * @Route("/api/news/{id}/delete", name="news_delete")
@@ -44,12 +76,11 @@ class NewsController extends AbstractController
 
         $news = $this->getDoctrine()->getRepository(News::class)->find($id);
         
-        $entityManager = $this->getDoctrine()->getManager();
-        $entityManager->remove($news);
-        $response = new Response();
-        $response->send();
-
-        return $this->json($news);
-      }
-
+            if ($news) {
+                $entityManager = $this->getDoctrine()->getManager();
+                $entityManager->remove($news);
+                $entityManager->flush();
+            }
+      return new Response(null, 204);
+    }
 }
